Enum email::Address

pub enum Address {
    Mailbox(Mailbox),
    Group(String, Vec<Mailbox>),
}

Represents an RFC 5322 Address

Variants

Mailbox

A "regular" email address

Group

A named group of mailboxes

Methods

impl Address

fn new_mailbox(address: String) -> Address

Shortcut function to make a new Mailbox with the given address [unstable]

fn new_mailbox_with_name(name: String, address: String) -> Address

Shortcut function to make a new Mailbox with the address and given-name [unstable]

fn new_group(name: String, mailboxes: Vec<Mailbox>) -> Address

Shortcut function to make a new Group with a collection of mailboxes [unstable]