Enum email::MimeMultipartTypeStable

pub enum MimeMultipartType {
    Mixed,
    Alternate,
    Digest,
    Parallel,
}

Marks the type of a multipart message

Variants

Mixed

Entries which are independent.

This value is the default.

As defined by Section 5.1.3 of RFC 2046

Alternate

Entries which are interchangeable, such that the system can choose whichever is "best" for its use.

As defined by Section 5.1.4 of RFC 2046

Digest

Entries are (typically) a collection of messages.

As defined by Section 5.1.5 of RFC 2046

Parallel

Entry order does not matter, and could be displayed simultaneously.

As defined by Section 5.1.6 of RFC 2046

Methods

impl MimeMultipartType

fn from_content_type(ct: MimeContentType) -> Option<MimeMultipartType>

Returns the appropriate MimeMultipartType for the given MimeContentType

fn to_content_type(&self) -> MimeContentType

Returns a MimeContentType that represents this multipart type.