Enum email::mimeheaders::MimeContentTransferEncodingStable

pub enum MimeContentTransferEncoding {
    Identity,
    QuotedPrintable,
    Base64,
}

Special header type for the Content-Transfer-Encoding header.

Variants

Identity

Message content is not encoded in any way.

QuotedPrintable

Content transfered using the quoted-printable encoding.

This encoding is defined in RFC 2045 Section 6.7

Base64

Content transfered as BASE64

This encoding is defined in RFC 2045 Section 6.8

Methods

impl MimeContentTransferEncoding

fn decode(&self, input: &String) -> Option<Vec<u8>>

Decode the input string with this transfer encoding.

Note that this will return a clone of the input's bytes if the transfer encoding is the Identity encoding.