Struct email::rfc2045::Rfc2045ParserStable

pub struct Rfc2045Parser<'s> {
    // some fields omitted
}

Parser over RFC 2045 style headers.

Things of the style value; param1=foo; param2="bar"

Methods

impl<'s> Rfc2045Parser<'s>

fn new(s: &str) -> Rfc2045Parser

Create a new parser over s

fn consume_all(&mut self) -> (String, HashMap<String, String>)

Consume up to all of the input into the value and a hashmap over parameters to values.