Struct email::HeaderUnstable

pub struct Header {
    pub name: String,
    // some fields omitted
}

Represents an RFC 822 Header

Fields

name

The name of this header

Methods

impl Header

fn new(name: String, value: String) -> Header

Creates a new Header for the given name and value

fn new_with_value<T: ToFoldedHeader>(name: String, value: T) -> Option<Header>

Creates a new Header for the given name and value, as converted through the ToHeader or ToFoldedHeader trait.

Returns None if the value failed to be converted.

fn get_value<T: FromHeader>(&self) -> Option<T>

Get the value represented by this header, as parsed into whichever type T