Struct email::HeaderMapUnstable

pub struct HeaderMap {
    // some fields omitted
}

A collection of Headers

Methods

impl HeaderMap

fn new() -> HeaderMap

fn insert(&mut self, header: Header)

Adds a header to the collection

fn iter(&self) -> HeaderIter

Get an Iterator over the collection of headers.

fn get(&self, name: String) -> Option<&Header>

Get the last value of the header with name

fn get_value<T: FromHeader>(&self, name: String) -> Option<T>

Get the last value of the header with name, as a decoded type.

fn len(&self) -> usize

Get the number of headers within this map.

fn find(&self, name: &String) -> Option<Vec<&Header>>

Find a list of headers of name, None if there are no headers with that name.