Trait email::ToHeader

pub trait ToHeader {
    pub fn to_header(value: Self) -> ParsingResult<String>;
}

Trait for converting from a Rust type into a Header value.

Required methods

pub fn to_header(value: Self) -> ParsingResult<String>

Turn the value into a String suitable for being used in a message header.

Returns None if the value cannot be stringified.

Loading content...