Crate email

Modules

mimeheaders

results

rfc2045
Module for dealing with RFC2045 style headers.

rfc2047
Module for decoding RFC 2047 strings

rfc5322
Module with helpers for dealing with RFC 5322

rfc822

Structs

Header
Represents an RFC 822 Header [unstable]

HeaderIter
[unstable]

HeaderMap
A collection of Headers [unstable]

Mailbox
Represents an RFC 5322 mailbox

MimeMessage
Represents a MIME message [unstable]

Enums

Address
Represents an RFC 5322 Address

MimeMultipartType
Marks the type of a multipart message

Traits

FromHeader
Trait for converting from RFC822 Header values into Rust types.

ToFoldedHeader
Trait for converting from a Rust time into a Header value that handles its own folding.

ToHeader
Trait for converting from a Rust type into a Header value.