Trait email::ToFoldedHeaderUnstable

pub trait ToFoldedHeader {
    fn to_folded_header(start_pos: usize, value: Self) -> Option<String>;
}

Trait for converting from a Rust time into a Header value that handles its own folding.

Be mindful that this trait does not mean that the value will not be folded later, rather that the type returns a value that should not be folded, given that the header value starts so far in to a line.

Required Methods

fn to_folded_header(start_pos: usize, value: Self) -> Option<String>